CachedTree: fall back to original tree in getChilds
The method shows a super weird behaviour in the case where order and
direction parameters are used. It doesn't seem to actually order stuff
according to the parameter. The both parameters are rarely used anyway,
though. So I decided to implement a fallback to the original tree
implementation if the params are used.